+++ src/parser/block/paragraph.ts
@@ -1,8 +1,9 @@
 ﻿import { ParagraphParser } from '../block';
 import { verify } from './util/verification';
-import { combine, loop } from '../../combinator';
+import { combine, subsequence, loop } from '../../combinator';
+import { reference } from './paragraph/reference';
 import { hashtag } from './paragraph/hashtag';
-import { inline } from '../inline';
+import { inline, InlineParser } from '../inline';
 import { squash } from '../squash';
 
 const separator = /^\s*$/m;
@@ -13,8 +14,12 @@ export const paragraph: ParagraphParser = verify((source: string) => {
   const block = source.split(separator, 1)[0];
   assert(block.length > 0);
   const rest = source.slice(block.length);
-  const [cs = []] = loop(combine<HTMLElement | Text, ParagraphParser.InnerParsers>([hashtag, inline]))(block.replace(emptyline, '').trim()) || [];
+  const [cs = []] = subsequence<HTMLElement | Text, ParagraphParser.InnerParsers>([loop(reference), loop(combine<HTMLElement | Text, [ParagraphParser.HashtagParser, InlineParser]>([hashtag, inline]))])(block.replace(emptyline, '').trim()) || [];
   const el = document.createElement('p');
   void el.appendChild(squash(cs));
+  void [...el.children]
+    .forEach(el =>
+      el.matches('.reference') && el.nextSibling &&
+      el.parentElement!.insertBefore(document.createElement('br'), el.nextSibling));
   return [[el], rest];
 });